Deleted sample directory
