+++ src/sve/gui/images/64-gazelle.jpg
