+++ src/sve/gui/images/cheetah2.png
