+++ src/sve/gui/images/img_tree_small2.png
